Updated to v5.3 beta.
Changes :
 * Added localization using external files. The program lists all .lng files present in its directory and relies on the '# langName' parameter to display them in the ComboBox at the bottom.
 * Added a specific folder in the SVN repository for language files. Only French and English at the moment.
 * Some settings are saved into 'settings.ini' and loaded when the program restarts : language (restart needed for the change to take effect) and game version.

TODO :
 * Localize messages (errors, warnings and informations). That will require some tweaking in the code.
 * Move the translation code into a separate file ; Form1.vb is getting too big.


 - About localization :

As I found no convenient way to use language files in .NET, I had to make a specific format.
It is based on INI format, but without sections.
Comments begin with '#' or ';'.
Some comments starting with '#' are special ; please do not remove the ones which are in the default language files.
You may edit anything that comes after the '=' character.
New lines (vbCrLf) are replaced by '|' ; use '\|' to insert a '|' and '\\|' to insert '\|' (that should not be needed but who knows :) ).
Leading and trailing spaces and tabulations are trimmed.

If no 'settings.ini' is in the program's directory, or if it is present but without 'Language File' setting, it tries to find a .lng file that matches the computer's locale, relying on the 2-char '# langID' parameter (e.g. 'en', 'fr'...)
If the text of a control is not present, it falls back to French for that control.
If there is no language file matching the computer's locale or if the file set by 'Language File' in the .ini doesn't exist, the whole program falls back to French.

To generate a default .lng file (texts as in the program), execute it with the command line argument '/generatelangfile'. The program will generate a 'default.lng' file and exit.

git-svn-id: svn://localhost/home/mat/projets/pokemon-signature-editor/pokemon-signature-editor_svn@16 98164a29-962a-4dce-8f92-5cef09f52806